measure transfer speed

+++ src/com/projectnine/logic/ITCPTransferActualizer.java
@@ -2,7 +2,7 @@ package com.projectnine.logic;
 
 public interface ITCPTransferActualizer {
 	public void setSinglePackageSize(int size);
-	public void setTotalPackageSize(long size);
-	public void setTotalTime(long seconds);
-	public void setSpeed(long seconds);
+	public void setTotalPackageSize(double size);
+	public void setTotalTime(double seconds);
+	public void setSpeed(double seconds);
 }
